module.exports & require to catch it
9.2.6
created a new JS to handle the generate HTML function and pasted it in.  I then used module.exports = generatePage (the funcs name) to send it out.  Then in the app.js I used const generatePage = require('./src/page-template.js'); to catch it and set it to the var generatePage.